move Card.component file to the cards directory
