+++ src/index.scss
@@ -8,5 +8,6 @@
 
 @import "/scss/components/button";
 @import "/scss/components/stats";
+@import "/scss/components/fullbgimage";
 
 @import "/scss/pages/home";
