+++ src/scss/base/_base.scss
@@ -26,9 +26,8 @@ html {
 
 body {
   box-sizing: border-box;
-  padding: 3rem;
+}
 
-  @include respond(tab-port) {
-    padding: 0;
-  }
+a {
+  text-decoration: none;
 }
